refactor: UserServiceImpl클래스) 메서드 수정
파라미터로 패스워드 받아서 검증하게끔 수정

+++ src/main/java/org/example/newsfeed/user/service/UserServiceImpl.java
@@ -95,11 +95,13 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, String password) {
 
         Users findUser = userRepository.findUserByIdOrElseThrow(id);
 
+        if (!passwordEncoder.matches(password,findUser.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"원래의 비밀번호가 일치하지 않습니다.");
+        }
         userRepository.delete(findUser);
-
     }
 }
